+++ color/make_x11_colors.py
@@ -239,12 +239,16 @@ if len(obuf) > 3:
 print("];")
 print()
 
-print(f"const NAMES: [&str; {n}] = [")
+print(f"pub(crate) const NAMES: [&str; {n}] = [")
 for (name, rgba) in keys:
     print(f'    "{name}",')
 print("];")
-print()
-print(f"const COLORS: [[u8; 4]; {n}] = [")
+print(f"""
+/// RGBA8 color components of the named X11 colors, in the same order as [`NAMES`].
+///
+/// Use [`lookup_palette_index`] to efficiently find the color components for a given color name
+/// string.
+pub(crate) const COLORS: [[u8; 4]; {n}] = [""")
 for (name, rgba) in keys:
     print(f'    {list(rgba)},')
 print("];")
@@ -260,7 +264,9 @@ fn weak_hash(key: u32, salt: u32, n: usize) -> usize {
     (((y as u64) * (n as u64)) >> 32) as usize
 }
 
-pub(crate) fn lookup_palette(s: &str) -> Option<[u8; 4]> {
+/// Given a named color (e.g., "red", "mediumorchid"), returns the index of that color into
+/// [`COLORS`] and [`NAMES`].
+pub(crate) fn lookup_palette_index(s: &str) -> Option<usize> {
     let mut key = 0_u32;
     for b in s.as_bytes() {
         key = key.wrapping_mul(9).wrapping_add(*b as u32);
@@ -268,7 +274,7 @@ pub(crate) fn lookup_palette(s: &str) -> Option<[u8; 4]> {
     let salt = SALTS[weak_hash(key, 0, SALTS.len())] as u32;
     let ix = weak_hash(key, salt, SALTS.len());
     if s == NAMES[ix] {
-        Some(COLORS[ix])
+        Some(ix)
     } else {
         None
     }