+++ color/src/lib.rs
@@ -86,8 +86,8 @@
 
 mod color;
 mod colorspace;
+mod flags;
 mod gradient;
-mod missing;
 // Note: this may become feature-gated; we'll decide this soon
 mod dynamic;
 pub mod palette;
@@ -109,8 +109,8 @@ pub use colorspace::{
     Oklch, ProphotoRgb, Rec2020, Srgb, XyzD50, XyzD65,
 };
 pub use dynamic::{DynamicColor, Interpolator};
+pub use flags::{Flags, Missing};
 pub use gradient::{gradient, GradientIter};
-pub use missing::Missing;
 pub use parse::{parse_color, parse_color_prefix, ParseError};
 pub use rgba8::{PremulRgba8, Rgba8};
 pub use tag::ColorSpaceTag;